Revert "fixing get_fps()"
that was bull. the error is elswhere and everywhere!!!

This reverts commit 3149fea657539f32146b4d84b8a9399070cc95cd.

git-svn-id: svn://dyne.org/rastasoft/freej/branches/Goil/freej_91@1083 383723c8-4afa-0310-b8a8-b1afb83214fc

+++ src/include/jsync.h
@@ -70,7 +70,7 @@ class JSyncThread {
   void signal_feed() { pthread_cond_signal(&_cond_feed); };
   int sleep_feed();
 
-  double get_fps();
+  float get_fps();
   void set_fps(float);
   bool running, quit;
